Restructures the utils folder
Moves all the helpers function to a newly created helpers folder and documents them with JSDoc format.
